float

+++ streamlit_app.py
@@ -7,7 +7,7 @@ import numpy as np
 def main():
   
   dr_temp_x = np.arange(-60,141)
-  dr_dr_y = np.arange(0.5,1.31, 0.01)
+  dr_dr_y = np.arange(0.5,1.31, 0.01, dtype=float)
   
   st.write(dr_dr_y)
